fix: .coveragerc .idea/modules.xml .idea/mongoclass.iml Makefile pyproject.toml pytest.ini ruff.toml mongoclass/__init__.py -> src/mongodata/__init__.py mongoclass/cache.py -> src/mongodata/cache.py mongoclass/client.py -> src/mongodata/client.py mongoclass/cursor.py -> src/mongodata/cursor.py tests/mongita/test_basic.py tests/mongita/test_cursor.py tests/mongita/test_insert.py tests/mongoclass/test_is_testing.py tests/mongoclass/test_mongodb.py tests/pymongo/test_basic.py tests/pymongo/test_cursor.py tests/utils.py tox.ini

+++ tests/utils.py
@@ -1,23 +1,24 @@
 from dataclasses import dataclass
 
-from mongoclass import client_constructor
+from mongodata import client_constructor
 
 from mongita import MongitaClientDisk
 from pymongo import MongoClient
 
 HOSTS = ["localhost:27017", "./mongita"]
-DATABASES = ["mongoclass", "coordinates_data", "profiles_list", "mongoclass-test"]
+DATABASES = ["mongoclass", "coordinates_data", "profiles_list", "mongoclass-test", "coordinates_data-test",
+             "profiles_list-test"]
 
 
 def drop_database(database: str = None) -> None:
-
     clients = [MongoClient(HOSTS[0]), MongitaClientDisk(HOSTS[1])]
     for client in clients:
         if database:
             client.drop_database(database or DATABASES[0])
         else:
             for d in DATABASES:
-                client.drop_database(d)
+                if d in client.list_database_names():
+                    client.drop_database(d)
 
 
 def create_client(engine: str = "pymongo"):
